@@ -1,7 +1,10 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/foundation.dart' show kIsWeb;
 
 void main() {
-  runApp(MaterialApp(title: "App Aula 01 - Contadores", home: Home()));
+  runApp(MaterialApp(
+      title: "Contador de Pessoas",
+      home: Home()));
 }
 
 class Home extends StatefulWidget {
@@ -10,8 +13,12 @@ class Home extends StatefulWidget {
 }
 
 class _HomeState extends State<Home> {
+
   int _people = 0;
   String _infoText = "Pode entrar!";
+
+  Color _baseColor = kIsWeb ? Colors.black : Colors.white;
+
   void _changePeople(int delta) {
     setState(() {
       _people += delta;
@@ -31,60 +38,59 @@ class _HomeState extends State<Home> {
 
   @override
   Widget build(BuildContext context) {
-    return Stack(children: <Widget>[
-      SizedBox.expand(
-          child: Image.asset(
-        "images/hotel.jpg",
-        color: Color.fromRGBO(255, 255, 255, 0.5),
-        colorBlendMode: BlendMode.modulate,
-        fit: BoxFit.fill,
-      )),
-      Padding(
-        padding: EdgeInsets.all(60.0),
-        child: Text("Hotel Tama",
-        style: TextStyle(color: Colors.white),)
-      ),
-      Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: <Widget>[
-          Text(
-            "Pessoas: $_people",
-            style: TextStyle(fontWeight: FontWeight.bold, color: Colors.white),
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              Padding(
-                padding: EdgeInsets.all(10.0),
-                child: FlatButton(
-                  child: Text("-1",
-                      style: TextStyle(fontSize: 40.0, color: Colors.white)),
-                  onPressed: () {
-                    _changePeople(-1);
-                  },
+    return Stack(
+      alignment: Alignment.center,
+      children: [
+        Image.asset("images/hotel.jpg",
+          fit: BoxFit.fitHeight,
+          color: Color.fromRGBO(255, 255, 255, 0.5),
+          colorBlendMode: BlendMode.modulate,
+        ),
+        Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            Padding(
+                padding: EdgeInsets.all(60.0),
+                child: Text("Hotel Tama",
+                  style: TextStyle(color: _baseColor),)
+            ),
+            Text(
+              "Pessoas: $_people",
+              style: TextStyle(color: _baseColor, fontWeight: FontWeight.bold),
+            ),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                Padding(padding: EdgeInsets.all(10.0),
+                  child:   TextButton(
+                    child: Text(
+                      "+1",
+                      style: TextStyle(fontSize: 40.0, color: _baseColor),
+                    ),
+                    onPressed: (){_changePeople(1);},
+                  ),
                 ),
-              ),
-              Padding(
-                padding: EdgeInsets.all(10.0),
-                child: FlatButton(
-                  child: Text("+1",
-                      style: TextStyle(fontSize: 40.0, color: Colors.white)),
-                  onPressed: () {
-                    _changePeople(1);
-                  },
+                Padding(padding: EdgeInsets.all(10.0),
+                  child:   TextButton(
+                    child: Text(
+                      "-1",
+                      style: TextStyle(fontSize: 40.0, color: _baseColor),
+                    ),
+                      onPressed: (){_changePeople(-1);}
+                  ),
                 ),
-              ),
-            ],
-          ),
-          Text(
-            _infoText,
-            style: TextStyle(
-                fontStyle: FontStyle.italic,
-                color: Colors.white,
-                fontSize: 15.0),
-          )
-        ],
-      ),
-    ]);
+              ],
+            ),
+            SelectableText(
+              _infoText,
+              style: TextStyle(
+                  color: _baseColor,
+                  fontStyle: FontStyle.italic,
+                  fontSize: 15.0),
+            ),
+          ],
+        )
+      ],
+    );
   }
 }
